Create WyLL target #803
This creates the WyLL backend target.  The intention with this is to
provide a general compilation target which is similar (in some sense) to
a typical compilation target (e.g. Java, JavaScript, C).  We can think
of WyLL as a generic imperative programming language.

This also adds the TypeMangler interface, along with the StdTypeMangler
implementation.  The latter is based upon the mangled that was being
used for WyJS.  At this stage, the StdTypeMangler does not correspond
exactly to its specification, and needs to be updated.