Create concept of a "low level driver".
A low level driver is something which compiles Whiley to a lower level
representation.  Some rought assumptions are currently made about what
is likely to be supported in the target architecture.  For example,
quantifiers and multiple returns are assumed not to be supported.

This includes support for aribtrary switch statements.  In some cases, these
are compiled down to low-level switch statements.  In others, they are
compiled down to ifelse chains.

Finally, this will effectively replace the WyLL target developed
previously.  In essence, it achieves the same thing with much less code
and abstraction.  It's just much simpler and potentially more easily
extended.