SemanticType now implements SyntacticType
This means that SemanticType is now properly in the WhileyFile
hierarchy, along with the other types and expressions.  This is helpful
because it allows e.g. visitors to be used with them seemlessly.  It
also means they could potentially be stored into a WyilFile if needed
(e.g. to support incremental computation).