FlowTypeChecker now employs type propagation
The FlowTypeChecker now employs a (slightly broken) form of type
propagation.  This propagates type information from the destination of
a given expression down the AST making up that expression.  This allows
one to determine the appropriate representation at each point.  In
addition, it means we can compute the definitive (concrete) type for
each expression.  Hopefully this solves a set of long standing issues
related to the translation of Whiley for low-level backends.  Fingers
crossed!!